% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tokens.R
\name{as_tokens}
\alias{as_tokens}
\title{Create a list of tokens}
\usage{
as_tokens(tbl, type, pos = TRUE, ...)
}
\arguments{
\item{tbl}{A data.frame of tokens out of \code{tokenize_to_df()}.}

\item{type}{Preference for the format of returned tokens.
Pick one of "surface", "dictionary", "normalized", or "reading".}

\item{pos}{When passed as \code{TRUE}, this function
uses the part-of-speech information as the name of the returned tokens.}

\item{...}{Passed to \code{dict_features()}.}
}
\value{
A named list of character vectors.
}
\description{
Create a list of tokens
}
\examples{
\dontrun{
tokenize_to_df("Tokyo, Japan") |>
  as_tokens(type = "surface")
}
}
