% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{create_sudachipy_env}
\alias{create_sudachipy_env}
\title{Create virtualenv env used by sudachir}
\usage{
create_sudachipy_env(python_version = "3.9.12")
}
\arguments{
\item{python_version}{Python version to use
within the virtualenv environment created.
SudachiPy requires Python 3.6 or higher to install.}
}
\description{
Create virtualenv env used by sudachir
}
