% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_features.R
\name{dict_features}
\alias{dict_features}
\title{Get dictionary's features}
\usage{
dict_features(lang = c("ja", "en"))
}
\arguments{
\item{lang}{Dictionary features label; one of "ja" or "en".}
}
\description{
Get dictionary's features
}
\examples{
dict_features("en")
}
