% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form}
\alias{form}
\title{Parse tokenized input text}
\usage{
form(x, mode, type)
}
\arguments{
\item{x}{Input text vectors}

\item{mode}{Select split mode (A, B, C)}

\item{type}{return form. "dictionary" or "normalized".}
}
\description{
Parse tokenized input text
}
