% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{remove_sudachipy}
\alias{remove_sudachipy}
\title{Remove SudachiPy}
\usage{
remove_sudachipy()
}
\description{
Uninstalls SudachiPy by removing the Conda environment.
}
\examples{
\dontrun{
install_sudachipy()
remove_sudachipy()
}
}
