% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{sudachir-defunct}
\alias{sudachir-defunct}
\alias{tokenizer}
\title{'Sudachi' tokenizer}
\usage{
tokenizer(...)
}
\arguments{
\item{...}{Not used.}
}
\description{
The old \code{tokenizer()} function was removed.
}
\details{
In general, users should not directly touch
the \verb{<sudachipy.tokenizer.Tokenizer>} and its \code{MorphemeList} objects.
If you must access those objects,
use the return value of the \code{rebuild_tokenizer()} function.
}
