% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize}
\alias{tokenize}
\title{Call 'Sudachi' tokenizer}
\usage{
tokenize(x, text_field, docid_field, instance)
}
\arguments{
\item{x}{A data.frame like object or a character vector to be tokenized.}

\item{text_field}{Column name where to get texts to be tokenized.}

\item{docid_field}{Column name where to get identifiers of texts.}

\item{instance}{A binding to the instance of \verb{<sudachipy.tokenizer.Tokenizer>}.
If you already have a tokenizer instance,
you can improve performance by providing a predefined instance.}
}
\description{
Call 'Sudachi' tokenizer
}
\keyword{internal}
