% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{as_tokens}
\alias{as_tokens}
\title{Call 'Sudachi' tokenizer}
\usage{
as_tokens(
  sentence,
  doc_id = seq_along(sentence),
  instance = rebuild_tokenizer()
)
}
\arguments{
\item{sentence}{Input text vectors.}

\item{doc_id}{Identifier of input sentences.}

\item{instance}{This is optional; if you already have an instance of
\verb{<sudachipy.tokenizer.Tokenizer>}, supplying the predefined
instance would improve the performance.}
}
\description{
Call 'Sudachi' tokenizer
}
\examples{
\dontrun{
as_tokens("Tokyo, Japan")
}
}
