% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form}
\alias{form}
\title{Create a list of tokens}
\usage{
form(tbl, type, pos = TRUE, ...)
}
\arguments{
\item{tbl}{A data.frame of tokens out of \code{tokenize_to_df}.}

\item{type}{Preference for the form of returned tokens.
Pick one of "surface", "dictionary", "normalized", "reading".}

\item{pos}{When supplied with \code{TRUE}, this function
uses the part-of-speech information as the name of the returned tokens.}

\item{...}{Passed to \code{dict_features}.}
}
\description{
Create a list of tokens
}
\examples{
\dontrun{
tokenize_to_df(
  "Tokyo, Japan",
  mode = "A"
) |>
form(type = "surface")
}
}
