% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_phrase.R
\name{form}
\alias{form}
\title{Create a list of tokens}
\usage{
form(
  sentence,
  doc_id = seq_along(sentence),
  mode = c("C", "B", "A"),
  into = dict_features(),
  col_select = seq_along(into),
  instance = NULL,
  ...
)
}
\arguments{
\item{sentence}{Input text vectors.}

\item{doc_id}{Identifier of input sentences.}

\item{mode}{Split mode (A, B, C)}

\item{into}{Column names of features.}

\item{col_select}{Character or integer vector of column names
that kept in the returned value.}

\item{instance}{This is optional; if you already have an instance of
\verb{<sudachipy.tokenizer.Tokenizer>}, supplying the predefined
instance would improve the performance.}

\item{...}{Passed to \code{as_phrase}.}
}
\description{
This function is a shorthand of \code{tokenize_to_df() |> as_phrase()}.
}
\examples{
\dontrun{
form(
  "Tokyo, Japan",
  mode = "A",
  type = "surface"
)
}
}
