% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_sudachipy}
\alias{install_sudachipy}
\title{Install SudachiPy}
\usage{
install_sudachipy()
}
\description{
Install SudachiPy to virtualenv virtual environment. As a one-time setup step,
you most run \code{install_sudachipy()} to install all dependencies.
}
\details{
\code{install_sudachipy()} requires Python and virtualenv to be installed.
See <https://www.python.org/getit/.
}
\examples{
\dontrun{
install_sudachipy()
}
}
