% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_sudachipy}
\alias{install_sudachipy}
\title{Install SudachiPy}
\usage{
install_sudachipy()
}
\description{
Install SudachiPy to Conda virtual environment. As a one-time setup step,
you most run \code{install_sudachipy()} to install all dependencies.
}
\details{
\code{install_sudachipy()} requires Python and Conda to be installed.
See \url{https://www.python.org/getit/} and
\url{https://docs.conda.io/projects/conda/en/latest/user-guide/install/}.
}
\examples{
\dontrun{
install_sudachipy()
}
}
