% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{rebuild_tokenizer}
\alias{rebuild_tokenizer}
\title{Rebuild 'Sudachi' tokenizer}
\usage{
rebuild_tokenizer(dict_type = c("core", "small", "full"), config_path = NULL)
}
\arguments{
\item{dict_type}{Dictionary type.}

\item{config_path}{Absolute path to \code{sudachi.json}}
}
\value{
Returns a binding to the instance of \verb{<sudachipy.tokenizer.Tokenizer>}.
}
\description{
Rebuild 'Sudachi' tokenizer
}
\examples{
\dontrun{
tokenizer <- rebuild_tokenizer()
as_tokens("Tokyo, Japan", instance = tokenizer)
}
}
