% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{rebuild_tokenizer}
\alias{rebuild_tokenizer}
\title{Rebuild tokenizer}
\usage{
rebuild_tokenizer(config_path = NULL)
}
\arguments{
\item{config_path}{Absolute path to \code{sudachi.json}}
}
\value{
Returns a binding to the instance of \verb{<sudachipy.tokenizer.Tokenizer>}.
}
\description{
Rebuild tokenizer
}
\examples{
\dontrun{
instance <- rebuild_tokenizer()
tokenizer("Tokyo, Japan", mode = "A", instance)
}
}
