% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{tokenize_to_df}
\alias{tokenize_to_df}
\title{Create tokenizing data.frame using Sudachi}
\usage{
tokenize_to_df(x, mode)
}
\arguments{
\item{x}{Input text vectors}

\item{mode}{Select split mode (A, B, C)}
}
\description{
Create tokenizing data.frame using Sudachi
}
